# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Doc",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header.\nA valid JWT token is generated by the login API and returned as response. The following syntax must be used \nin the 'Authorization' header\nJWT xxxxxxxx_yyyy-zzzzzzz\n"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "paths": {
    "/authenticate": {
      "post": {
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Login",
        "description": "Login for registered user\n",
        "tags": [
          "User Authentication"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Login Credentials",
            "required": true,
            "description": "User Email and password",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login Successfull",
            "schema": {
              "$ref": "#/definitions/loginSuccess"
            }
          },
          "500": {
            "description": "Something went wrong"
          }
        }
      }
    },
...
    "/randomatSign": {
      "get": {
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "random atSign",
        "description": "random atSign user\n",
        "tags": [
          "User list"
        ],
        "responses": {
          "200": {
            "description": "Invite Sent",
            "schema": {
              "$ref": "#/definitions/randomatSignResponse"
            }
          },
          "500": {
            "description": "Something went wrong"
          }
        }
      }
    },
...
  },
  "definitions": {
    "login": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "name": "email",
          "type": "string",
          "description": "a valid email id"
        },
        "password": {
          "name": "email",
          "type": "string",
          "description": "user password"
        }
      }
    },
...
    "loginSuccess": {
      "example": {
        "user": {
          "mobileVerified": false,
          "reserveTime": "1582391138912",
          "_id": "5e331fdd8f98df04e4c50b6c",
          "invitedOn": "2020-01-30T18:26:37.184Z",
          "email": "a valid email",
          "contact": "0090",
          "userRole": "Admin",
          "userStatus": "Active",
          "password": "$2a$10$At6LlTVsuWYc68JU.YsSYuIuEHNFYHF48rqiNytqEFRZIlvU16Mmy",
          "saltSecret": "$2a$10$l3OIAouyRmMDp02gPEjgKe",
          "__v": 0
        },
        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI1ZTMzMWZkZDhmOThkZjA0ZTRjNTBiNmMiLCJpYXQiOjE1ODIzOTExMzksImV4cCI6MTU4NTk5MTEzOX0.PxHc4z5Y9bFReE4SNa7Eb-GfJBzJ9__FUzR7CuaEUHI"
      }
    },
...
    "randomatSignResponse": {
      "example": {
        "status": "success",
        "message": "",
        "data": [
          "your",
          "everythingisilluminatedicy"
        ]
      }
    },
...
  }
}